refactor: pass project to adapter module
