+++ packages/iceworks-server/src/interface/base.ts
@@ -4,18 +4,15 @@
 export interface IProject {
   readonly name: string;
   readonly path: string;
-  readonly packageJSON: any;
-  readonly processEnv: any;
+  getPackageJSON(): any;
+  getEnv(): any;
 }
 
 /**
  * 功能模块的基类
  */
 export interface IBaseModule {
-  readonly projectName: string;
-  readonly projectPath: string;
-  readonly projectPackageJSON: any;
-  readonly processEnv: any;
+  project: IProject;
 }
 
 /**
